requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
psycopg2-binary==2.9.10
psycopg2==2.9.10
wheel==0.45.0
WTForms==3.2.1
Flask-WTF==1.2.2
pillow==11.1.0
bs4==0.0.2
pandas==2.2.3
email-validator==2.2.0
python-dotenv==1.0.1
gunicorn==20.1.0
